last step : norme && push

+++ fdf.h
@@ -6,7 +6,7 @@
 /*   By: absaid <absaid@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/27 08:34:16 by absaid            #+#    #+#             */
-/*   Updated: 2022/12/11 01:56:06 by absaid           ###   ########.fr       */
+/*   Updated: 2022/12/11 03:24:47 by absaid           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,9 +55,9 @@ typedef struct s_map
 	struct s_map *next;
 } t_map;
 
-t_point rot_x(t_point p, int angle);
-t_point rot_y(t_point p, int angle);
-t_point rot_z(t_point p, int angle);
+t_point rot_x(t_point p, double angle);
+t_point rot_y(t_point p, double angle);
+t_point rot_z(t_point p, double angle);
 t_point	iso(t_point p);
 int	ft_mapsize(t_map *lst);
 int ***get_coords(t_map *lst);
